fix miller-rabin and dijkstra

+++ snippets/graphs/dijkstra.cpp
@@ -2,7 +2,7 @@
 struct Edge{ int v; ll weight; };	// input edges
 struct PQ{ ll d; int v; };			// distance and target
 bool operator>(const PQ &l, const PQ &r){ return l.d > r.d; }
-int dijkstra(vector<vector<Edge>> &edges, int s, int t) {
+ll dijkstra(vector<vector<Edge>> &edges, int s, int t) {
 	vector<ll> dist(edges.size(),LLINF);
 	priority_queue<PQ,vector<PQ>,greater<PQ>> pq;
 	dist[s] = 0; pq.push({0, s});
